deleted unused files and refactored frontend
