+++ frontend/src/Components/Navbar.jsx
@@ -1,10 +1,7 @@
 import React, {useState, useEffect} from 'react';
-import { BrowserRouter as useParams } from 'react-router-dom';
-import useFetch from '../useFetch';
-import { AppBar, Toolbar, Typography, Button, Box } from '@mui/material';
-import { Link, Outlet, useLocation } from 'react-router-dom';
+import { AppBar, Toolbar, Typography } from '@mui/material';
+import { Link } from 'react-router-dom';
 import AccountCircleIcon from '@mui/icons-material/AccountCircle';
-import MenuIcon from '@mui/icons-material/Menu';
 import { IconButton } from '@mui/material';
 
 const Navbar = ({ isLoggedIn, onSignOut, projectId }) => {
@@ -25,22 +22,15 @@ const Navbar = ({ isLoggedIn, onSignOut, projectId }) => {
 		setName(data);
     })
     .catch(error => {
-     	console.error('Error', error);
+      console.error('Error', error);
     });
   }, []);
-
-  // const textOutline = `
-  //   -1px -1px 0 #000, 
-  //   1px -1px 0 #000, 
-  //   -1px 1px 0 #000, 
-  //   1px 1px 0 #000`; 
 
   return (
     <div>
       <AppBar position="fixed" sx={{
           backgroundColor: '#3c88bf',
           padding: '8px 16px',
-          // boxShadow: '1px 5px 10px rgba(0, 0, 0, 0.7)'
         }}>
         <Toolbar 
           disableGutters 
@@ -55,12 +45,14 @@ const Navbar = ({ isLoggedIn, onSignOut, projectId }) => {
               marginRight: '30px',
             }}
           >
-            {/* ScrumMate Logo */}
+            {/* ScrumMate Text when clicked, links to dashboard */}
             <Link exact to="/" style={{ textDecoration: 'none', color: 'white' }}>ScrumMate</Link>
           </Typography>
-              <Typography>
-                {name}
-              </Typography>
+
+          <Typography>
+            {name}
+          </Typography>
+          
           {/* Avatar Button */}
           {isLoggedIn && 
             <IconButton
