+++ frontend/src/Pages/ReleasePlan.jsx
@@ -1,13 +1,13 @@
 import React, { useState, useEffect } from 'react';
-import { Typography, Button, Box, Paper } from '@mui/material';
-import { List, ListItem } from '@mui/material';
-import { Grid, Input, Divider, Card, CardContent } from '@mui/material';
+import { Typography, Box } from '@mui/material';
+import { Grid, Divider } from '@mui/material';
 import Sidebar from '../Components/Sidebar';
 import ButtonBar from '../Components/ButtonBar';
 import ContentBox from '../Components/ContentBox';
-import Sprint from '../Components/Sprint';
-import { Dialog, DialogTitle, DialogContent, DialogActions, TextField, Select, MenuItem, FormControl, InputLabel } from '@mui/material';
 import DragList from '../Components/DragList';
+import Backlog from '../Components/Backlog';
+import SanityCheckGraph from '../Components/SanityCheckGraph';
+import SanityCheckText from '../Components/SanityCheckText';
 
 const ReleasePlan = () => {
   const [sprints, setSprints] = useState([]);
@@ -15,56 +15,51 @@ const ReleasePlan = () => {
   const [problemStatement, setProblem] = useState("");
   const [highLevelGoals, setGoals] = useState("");
   const [releaseId, setId] = useState(1);
-  const [backlogItems, setBacklogItems] = useState([]);
-  const [newBacklogType, setNewBacklogType] = useState('story');
-  const [dialogOpen, setDialogOpen] = useState(false);
-  const [newBacklogDescription, setNewBacklogDescription] = useState('');
-
 
 	const projectId = 1;
-	
+
 	function fetchMostRecentRelease(projectId, setProblem, setGoals, setId) {
-		console.log("about to most recent release")
+		console.log("about to most recent release");
 		var options = {
 			method:'get',
 			credentials:'include'
-		  }
+    };
 		fetch(`http://localhost:8080/api/project/${projectId}/recentRelease`, options).then((result)=>{
 			if(result.status === 200){
-				console.log(result)
+				console.log(result);
 			}
 			result.json().then((response)=>{
-				console.log(response)
-				setProblem(response.problemStatement)
-				setGoals(response.goalStatement)
-				setId(response.id)
+				console.log(response);
+				setProblem(response.problemStatement);
+				setGoals(response.goalStatement);
+				setId(response.id);
 			})
 		})
 	}
 
 	function fetchRelease(releaseId, setProblem, setGoals) {
-		console.log("about to fetch a release")
+		console.log("about to fetch a release");
 		var options = {
 			method:'get',
 			credentials:'include'
-		  }
+    };
 		fetch(`http://localhost:8080/api/release/${releaseId}`, options).then((result)=>{
 			if(result.status === 200){
 				console.log(result)
 			}
 			result.json().then((response)=>{
-				console.log(response)
-				setProblem(response.problemStatement)
-				setGoals(response.goalStatement)
-			})
-		})
+				console.log(response);
+				setProblem(response.problemStatement);
+				setGoals(response.goalStatement);
+			});
+		});
 	}
 
   function fetchSprints(releaseId) {
 		var options = {
 			method:'get',
 			credentials:'include'
-		  }
+    }
 		fetch(`http://localhost:8080/api/release/${releaseId}/sprints`, options).then((result)=>{
 			if(result.status === 200){
 				result.json().then((response)=>{
@@ -72,7 +67,7 @@ const ReleasePlan = () => {
 			} else {
 				setSprints([]);
 			}
-			});
+    });
   }
 
   useEffect(() => {
@@ -81,35 +76,14 @@ const ReleasePlan = () => {
 
   useEffect(() => {
     fetchRelease(releaseId, setProblem, setGoals);
-    fetchSprints(releaseId)
+    fetchSprints(releaseId);
   }, [releaseId]);
 
   const toggleDrawer = () => {
     setOpen(!open);
   };
 
-  const addBacklogItem = () => {
-    setDialogOpen(true); // This triggers the dialog to open
-  };
-
-  const handleAddBacklogItem = () => {
-    const newBacklogItems = [...backlogItems, { type: newBacklogType, description: newBacklogDescription }];
-    setBacklogItems(newBacklogItems);
-    setDialogOpen(false); // Close the dialog
-    setNewBacklogDescription(''); // Reset for the next item
-    setNewBacklogType('story'); // Reset the type for the next item
-  };
-
-  //data for sanity check
-  const barData = [
-    {pointsCompleted: 20 },
-    {pointsCompleted: 30 },
-    {pointsCompleted: 40 },
-    {pointsCompleted: 35 },
-    {pointsCompleted: 45 },
-  ];
-
-  // TODO: update Placeholder functions/variables with actual data
+  // TODO: update Placeholder functions with actual data
   const sprintPlanClick = () => console.log('Clicked Sprint Plan');
   const scrumBoardClick = () => console.log('Clicked Scrum Board');
   const burnupChartClick = () => console.log('Clicked Burnup Chart');
@@ -194,11 +168,9 @@ const ReleasePlan = () => {
         </Typography>
 
         {/* Problem Statement */}
-        {/* TODO: replace problem statement */}
         <ContentBox title={'Problem Statement'} content={problemStatement} />
 
         {/* High Level Goals */}
-        {/* TODO: high level goals */}
         <ContentBox title={'High Level Goals'} content={highLevelGoals} />
 
         <Grid container spacing={2}>
@@ -217,133 +189,31 @@ const ReleasePlan = () => {
             {/* {sprints != [] ? <DragList items={sprints} setItems={setSprints}/>: ''} */}
           </Grid>
 
+          {/* Backlog */}
           <Grid item xs={3}>
-            <Typography
-              variant="h6"
-              marginLeft={2}
-              textAlign={'left'}
-              fontWeight="bold"
-              fontSize={14}
-            >
-              Backlog
-            </Typography>
-
-            <Paper
-              sx={{
-                maxWidth: '90%',
-                marginLeft: 2,
-                backgroundColor: 'lightgray',
-              }}
-            >
-              <List>
-                {backlogItems.map((item, index) => (
-                  <ListItem key={index}>
-                    <Card
-                      sx={{
-                        marginBottom: 1,
-                      }}
-                    >
-                      <CardContent>
-                        <Typography
-                          variant="body1"
-                          fontSize={14}
-                        >
-                          <Input
-                            type="text"
-                            value={item.description}
-                            onChange={(e) => {
-                              const newBacklogItems = [...backlogItems];
-                              newBacklogItems[index].description = e.target.value;
-                              setBacklogItems(newBacklogItems);
-                            }}
-                            placeholder="Enter backlog item"
-                            style={{ border: 'none', width: '100%', padding: '4px' }}
-                          />
-                        </Typography>
-                      </CardContent>
-                    </Card>
-                  </ListItem>
-                ))}
-              </List>
-              {/* Button to add new backlog item */}
-              <Button
-                variant="contained"
-                onClick={addBacklogItem}
-                sx={{
-                  bgcolor: 'grey',
-                    '&:hover': {
-                      bgcolor: 'darkgrey', // Background color on hover
-                    },
-                  }}
-                >
-                Add Backlog Item +
-              </Button>
-            </Paper>
+            <Backlog />
           </Grid>
         </Grid>
 
-        {/* Sanity Checks */}
+        {/* Sanity Check */}
         <Typography variant="h5" align="left" fontWeight="bold" gutterBottom>
           Sanity Check
         </Typography>
+
         <Grid container spacing={2}>
+          {/* Sanity Check Graph */}
           <Grid item xs={12} sm={4}>
-            <Paper elevation={3} style={{ padding: '20px', marginTop: '20px' }}>
-              <Box display="flex" flexDirection="column" alignItems="flex-start">
-                {barData.map((data, index) => (
-                  <Box key={index} display="flex" alignItems="center" marginBottom={1}>
-                    <Box width={data.pointsCompleted * 5} height={30} bgcolor="darkgrey" />
-                    <Typography marginLeft={2} variant="body2">
-                      {data.pointsCompleted}
-                    </Typography>
-                  </Box>
-                ))}
-              </Box>
-            </Paper>
+            <SanityCheckGraph />
           </Grid>
 
           <Grid item xs={12} sm={7}> 
-            <Paper elevation={3} style={{ padding: '20px', marginTop: '20px', backgroundColor: 'lightgray', height: '200px' }}>
-              <Typography variant="body1" align="left">
-                Yes we can do it because no sprint looks like too much work. Lorem ipsum dolor sit amet …
-              </Typography>
-            </Paper>
+            <SanityCheckText 
+              text={'Yes we can do it because no sprint looks like too much work. Lorem ipsum dolor sit amet …'}
+            />
           </Grid>
         </Grid>
+
       </Grid>
-      <Dialog open={dialogOpen} onClose={() => setDialogOpen(false)}>
-        <DialogTitle>Add New</DialogTitle>
-        <DialogContent>
-          <FormControl fullWidth margin="normal">
-            <InputLabel>Type</InputLabel>
-            <Select
-              value={newBacklogType}
-              label="Type"
-              onChange={(e) => setNewBacklogType(e.target.value)}
-            >
-              <MenuItem value="story">Story</MenuItem>
-              <MenuItem value="spike">Spike</MenuItem>
-              <MenuItem value="infrastructure">Infrastructure</MenuItem>
-            </Select>
-          </FormControl>
-          <TextField
-            margin="dense"
-            id="description"
-            label="As a user..."
-            type="text"
-            fullWidth
-            multiline
-            rows={4}
-            variant="outlined"
-            value={newBacklogDescription}
-            onChange={(e) => setNewBacklogDescription(e.target.value)}
-          />
-        </DialogContent>
-        <DialogActions>
-          <Button onClick={() => setDialogOpen(false)}>Cancel</Button>
-          <Button onClick={handleAddBacklogItem}>Add Item</Button>
-        </DialogActions>
-      </Dialog>
     </Grid>
   );
 };
